修改base类，将toolbar转至fragment
