+++ app/src/main/java/us/xingkong/jueqian/base/BaseFragment.java
@@ -4,12 +4,16 @@ import android.app.ProgressDialog;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.umeng.analytics.MobclickAgent;
 
+import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import us.xingkong.jueqian.JueQianAPP;
@@ -24,10 +28,15 @@ import us.xingkong.jueqian.utils.ToastUtils;
 public abstract class BaseFragment<P extends BasePresenter> extends Fragment
         implements BaseView<P> {
 
+    @BindView(R.id.toolbar)
+    Toolbar mToolbar;
+
     private Unbinder mUnbinder;
     protected P mPresenter;
 
     protected ProgressDialog mProgressDialog;
+
+    private ActionBar mActionbar;
 
     @Nullable
     @Override
@@ -39,6 +48,8 @@ public abstract class BaseFragment<P extends BasePresenter> extends Fragment
         View root = inflater.inflate(bindLayout(), container, false);
         mUnbinder = ButterKnife.bind(this, root);
         prepareData(savedInstanceState);
+        initToolbar();
+        // 初始化视图
         initView(root);
         initData(savedInstanceState);
         initEvent();
@@ -86,6 +97,38 @@ public abstract class BaseFragment<P extends BasePresenter> extends Fragment
      * 初始化事件监听器
      */
     protected abstract void initEvent();
+
+
+    /**
+     * 初始化Toolbar
+     */
+    private void initToolbar() {
+        ((AppCompatActivity) getActivity()).setSupportActionBar(mToolbar);
+        mActionbar = ((AppCompatActivity) getActivity()).getSupportActionBar();
+    }
+
+    /**
+     * 设置Toolbar标题
+     *
+     * @param title 标题
+     */
+    protected void setToolbarTitle(String title) {
+        if (mActionbar != null) {
+            mActionbar.setTitle(title);
+        }
+    }
+
+    /**
+     * 设置Toolbar显示返回按钮及标题
+     *
+     * @param title 标题
+     */
+    protected void setToolbarBackEnable(String title) {
+        if (mActionbar != null) {
+            mActionbar.setDisplayHomeAsUpEnabled(true);
+            mActionbar.setHomeAsUpIndicator(R.drawable.ic_arrow_white_24dp);
+        }
+    }
 
     @Override
     public void showToast(CharSequence msg) {
