+++ app/src/main/res/layout/activity_main.xml
@@ -7,8 +7,6 @@
     android:orientation="vertical"
     tools:context="us.xingkong.jueqian.module.main.MainActivity">
 
-    <include layout="@layout/view_toolbar" />
-
     <us.xingkong.jueqian.widget.ScrollViewPager
         android:id="@+id/viewpager"
         android:layout_width="match_parent"
