+++ app/src/main/res/layout/fragment_reals_content.xml
@@ -2,9 +2,12 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/colorAccent"
     android:orientation="vertical">
 
-    <include layout="@layout/view_toolbar" />
+    <include
+        layout="@layout/view_toolbar"
+        android:visibility="gone" />
 
     <TextView
         android:id="@+id/tv_page_count"
